refactor(portal): split DynamicApplicationForm into focused hooks and pieces
DynamicApplicationForm was ~560 lines blending submit/draft/fee mutation,
companions state seeding, fee notice markup, merged-sections memoization
and the render loop. Pulls out three small, single-purpose modules:

- useSubmitApplication: owns the create/update mutation, the submit and
  save-draft handlers, and the Stripe fee redirect. Exposes the pending
  flags the UI needs.
- useCompanionsState: seeds companions from an existing application's
  attendees.
- ApplicationFeeNotice: the Info-icon + fee amount callout.

The main component now focuses on wiring and rendering. No behavior
change.
